+++ assets/linux/build.sh
@@ -34,7 +34,7 @@ dh_make --createorig --copyright gpl3 -e isaac@minnekdigital.com --single -y
 echo "cominnek /usr/bin" > debian/cominnek.install
 
 sed -i -e "s#<insert the upstream URL, if relevant>#${REPO_URL}#g" debian/control
-sed -i -e 's/Build-Depends: debhelper-compat (= 13)/&,git,git-flow/g' debian/control
+sed -i -e 's/Build-Depends: debhelper-compat (= 13)/&,git/g' debian/control
 sed -i -e 's/<insert up to 60 chars description>/Create commits and pull requests in an easy way/g' debian/control
 sed -i -e "s/ <insert long description, indented with spaces>/License: GPL-3.0/g" debian/control