+++ controllers/bridge/Download.go
@@ -16,19 +16,19 @@ func getRequest(downloadUrl string) *http.Request {
 	req, err := http.NewRequest("GET", downloadUrl, nil)
 
 	if err != nil {
-		log.Fatalln("ERR1: Someting goes wrong downloading the file")
+		log.Fatalln("ERR1: Something goes wrong downloading the file")
 	}
 
 	return req
 }
 
 func download(destinationPath, downloadUrl, fileName string) error {
-	tempDestinationPath := destinationPath + ".tmp"
+	tempDestinationPath := destinationPath
 
 	resp, err := http.DefaultClient.Do(getRequest(downloadUrl))
 
 	if err != nil {
-		log.Fatalln("ERR2: Someting goes wrong downloading the file")
+		log.Fatalln("ERR2: Something goes wrong downloading the file")
 	}
 
 	defer resp.Body.Close()
@@ -47,8 +47,8 @@ func download(destinationPath, downloadUrl, fileName string) error {
 	return nil
 }
 
-func DownloadFromURL(url, fileName string) string {
-	route := filepath.Join(config.Public.TempPath, fileName)
+func DownloadTempFromURL(url, fileName string) string {
+	route := filepath.Join(config.Public.TempPath, fileName+".tmp")
 
 	if !folders.CheckExists(config.Public.TempPath) {
 		folders.Create(config.Public.TempPath)
@@ -59,4 +59,35 @@ func DownloadFromURL(url, fileName string) string {
 	println()
 	log.Println("🎉 Downloaded")
 	return route
+}
+
+func DownloadFromURL(url, fileName, path string) string {
+	route := filepath.Join(path, fileName)
+
+	if !folders.CheckExists(path) {
+		folders.Create(path)
+	}
+
+	err := download(route, url, fileName)
+
+	if err != nil {
+		log.Fatalln("ERR3: Something goes wrong downloading the file")
+		return ""
+	}
+
+	println()
+	log.Println("🎉 Downloaded")
+	return route
+}
+
+func CheckIfExist(url string) bool {
+	resp, err := http.DefaultClient.Do(getRequest(url))
+
+	if err != nil {
+		log.Fatalln("ERR: Something goes wrong downloading the file")
+	}
+
+	defer resp.Body.Close()
+
+	return resp.StatusCode == 200
 }