+++ pkg/git/Commit.go
@@ -92,6 +92,46 @@ func GetAllCommits() []string {
 	return strings.Split(out, "\n")
 }
 
+func GetAllCommitsForRelease(lastReleaseHash string) []string {
+	cmd := "git log --pretty=format:'message[%s];hash[%h]'"
+
+	if lastReleaseHash != "" {
+		cmd += " " + lastReleaseHash + "..HEAD"
+	}
+
+	out, _, err := shell.Out(cmd)
+
+	if err != nil {
+		println("Error getting commits")
+		os.Exit(1)
+	}
+
+	return strings.Split(out, "\n")
+}
+
+func LastTag() string {
+	tags := GetTags()
+
+	for tag := range tags {
+		if !strings.Contains(tags[tag], "-") {
+			return tags[tag]
+		}
+	}
+
+	return ""
+}
+
+func GetTags() []string {
+	out, _, err := shell.Out("git tag --sort=-creatordate")
+
+	if err != nil {
+		println("Error getting tags")
+		return []string{}
+	}
+
+	return strings.Split(out, "\n")
+}
+
 func GetCommitHash(msg string) string {
 	return strings.Split(msg, ":")[0]
 }
@@ -110,7 +150,7 @@ func GetCommitByHash(hash string) string {
 		grepCmd = "Select-String"
 	}
 
-	cmd := fmt.Sprintf("git log --pretty=oneline --pretty=format:'%%h: %%s' %s | " + grepCmd + " %s", hash, hash)
+	cmd := fmt.Sprintf("git log --pretty=oneline --pretty=format:'%%h: %%s' %s | "+grepCmd+" %s", hash, hash)
 	out, _, err := shell.Out(cmd)
 
 	if err != nil {