+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [4.0.0](https://github.com/minnek-digital-studio/cominnek/compare/v3.1.0...4.0.0) (2023-08-07)
+
+### Features
+- feat: 90 add support to read project PR template ([53a43f4](https://github.com/minnek-digital-studio/cominnek/commit/53a43f4))
+- feat: add mnk-config.json to project ([55d9b13](https://github.com/minnek-digital-studio/cominnek/commit/55d9b13))
+- feat: add support to use flow package ([cb44a26](https://github.com/minnek-digital-studio/cominnek/commit/cb44a26))
+- feat: add support to download package ([8e91b49](https://github.com/minnek-digital-studio/cominnek/commit/8e91b49))
+- feat: 90 create support to custom flow ([1794664](https://github.com/minnek-digital-studio/cominnek/commit/1794664))
+- **branch:** 84 add sync branch ([50b3ebe](https://github.com/minnek-digital-studio/cominnek/commit/50b3ebe))
+
+### Bug Fixes
+- **publish:** 83 skip commit flag ([81d5fe5](https://github.com/minnek-digital-studio/cominnek/commit/81d5fe5))
+
 # [3.1.0] - 2023-04-30
 
 ## Added